% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VCI_earthquakes_plus}
\alias{VCI_earthquakes_plus}
\title{Example data for the function \code{\link{simulate_background_earthquakes}}}
\format{
An object of class data.frame with 7 columns,
formatted like the 7 columns of \bold{earthquake_data_plus},
the argument of the function \code{\link{simulate_background_earthquakes}}.
The magnitudes are moment magnitudes, as in \code{\link{VCI_earthquakes}}.
}
\usage{
VCI_earthquakes_plus
}
\description{
This catalog contains the earthquakes from
 \code{\link{VCI_earthquakes}}, for which the magnitude is greater than or equal to
 3.5 or the magnitude threshold that has been chosen.
}
\keyword{datasets}
