% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VCI_magnitude_sample}
\alias{VCI_magnitude_sample}
\title{First example data for the function \code{\link{simulate_aftershocks}}}
\format{
A numerical vector
}
\usage{
VCI_magnitude_sample
}
\description{
A data set containing only the moment magnitudes from \code{\link{VCI_earthquakes}}.
}
\keyword{datasets}
