% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VCI_simulated_background_earthquakes}
\alias{VCI_simulated_background_earthquakes}
\title{Second example data for the function
 \code{\link{simulate_aftershocks}}}
\format{
An object of class data.frame with 5 columns,
formatted like other data frames with the same structure described above,
including \code{\link{VCI_earthquakes}}.
}
\usage{
VCI_simulated_background_earthquakes
}
\description{
This earthquake data catalog is used as
 \bold{background_catalog} in an example of application of the function
 \code{\link{simulate_aftershocks}} and
 was obtained by running the
 example for the function \code{\link{simulate_background_earthquakes}}.
}
\keyword{datasets}
